Refactor duplicated code; introduce basic Meteor 1.4 file structure
