+++ imports/utilities/client/templates.js
@@ -1,0 +1,5 @@
+import { Template } from 'meteor/templating';
+
+const tmpl = () => Template.instance();
+
+export { tmpl };
